\name{hello}
\alias{hello}
\title{Hello, World!}
\usage{
hello(a = "a")
}
\description{
Prints 'Hello, world!' and the character given to the function, and prints
\code{'sausage'} returned from a nested C++ function.
}
\arguments{
  \item{a}{
character to be printed. Defaults to "a"
}
}
\examples{
hello("b")
}
